Finish refactor of frontpage tests
