+++ cypress/support/pageTestFragmentCommands.js
@@ -1,11 +1,35 @@
-Cypress.Commands.add('headerTest', () => {
+Cypress.Commands.add('headerTestBBCNewsString', () => {
   cy.get('header a').should('contain', 'BBC News');
 });
 
+Cypress.Commands.add('headerTestVisibleBanner', () => {
+  const el = {
+    viewport: 'viewport',
+    header: 'header',
+    section: 'section',
+  };
+  cy.get(el.header)
+    .should('have.lengthOf', 1)
+    .find('a')
+    .should('have.attr', 'href', 'https://www.bbc.co.uk/news') // unhardcode later
+    .find('svg')
+    .should('be.visible');
+});
+
+Cypress.Commands.add('headerTestHaveH1', () => {
+  cy.get('h1').should('have.length', 1);
+});
+
 Cypress.Commands.add('footerTestBranding', () => {
-  cy.get('footer a')
+  cy.get('footer')
+    .should('have.length', 1)
+    .should('have.attr', 'role', 'contentinfo')
+    .find('a')
     .eq(0)
-    .should('contain', 'BBC News');
+    .should('contain', 'BBC News')
+    .and('have.attr', 'href', 'https://www.bbc.co.uk/news') // unhardcode later
+    .find('svg')
+    .should('be.visible');
 });
 
 Cypress.Commands.add('footerTestWorkingLinks', () => {
@@ -30,4 +54,54 @@ Cypress.Commands.add('footerTestCopyrightLink', () => {
   cy.get('footer a')
     .eq(0)
     .should('contain', 'BBC News');
+});
+
+Cypress.Commands.add('hasScriptToFetchBundle', () => {
+  // Testing the actual fetch is not currently possible
+  cy.get('script')
+    .last()
+    .should('have.attr', 'src')
+    .and('match', /(\/static\/js\/main-\w+\.\w+\.js)/g);
+});
+
+Cypress.Commands.add('hasVisibleSectionLabel', () => {
+  const el = {
+    viewport: 'viewport',
+    header: 'header',
+    section: 'section',
+  };
+
+  cy.get(el.section)
+    .should('have.length.of.at.least', 1)
+    .should('be.visible')
+    .each($section => {
+      cy.wrap($section).within(() => {
+        cy.get('h2').should('have.lengthOf', 1);
+      });
+    });
+});
+
+Cypress.Commands.add('hasOneOrMoreStoryPromos', () => {
+  const el = {
+    viewport: 'viewport',
+    header: 'header',
+    section: 'section',
+  };
+
+  cy.get(el.section).within(() => {
+    cy.get('img')
+      .should('have.length.of.at.least', 1)
+      .should('be.visible');
+    cy.get('h3')
+      .should('have.length.of.at.least', 1)
+      .should('be.visible')
+      .find('a')
+      .should('have.attr', 'href');
+    cy.get('p')
+      .should('have.length.of.at.least', 1)
+      .should('be.visible');
+    cy.get('time')
+      .should('have.length.of.at.least', 1)
+      .should('be.visible');
+  });
 });
